+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- added mssql test server with seeding for kvis, ifsg and grippeweb
+
 ### Changes
 
 - use assets connector for test data
@@ -16,6 +18,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Deprecated
 
 ### Removed
+
+- BREAKING: remove support for python 3.11, 3.12, and 3.13
 
 ### Fixed